separate directories and files for each tab
